Question 1
Create a function called readNumbersFromFile that receives as a parameter a string with the name of a file (located in the same directory) of numbers. The file has a number on each line (assume all integers).

The function will open the file, read the values and then print (that is right, this function returns no values) the following:

The total of the values
The number of values (same as number of lines btw)
The arithmetic mean (average) of the values
The standard deviation of the values.